Extract common logic in Parquet and Feather to generic Arrow source.
